Moved timespan_from_outer_bounds from pipeline.provenance.util to pipeline.util.
