+++ pipeline/util/__init__.py
@@ -4,6 +4,8 @@ import sys
 import fnmatch
 import pprint
 import calendar
+import datetime
+import dateutil.parser
 from threading import Lock
 from contextlib import ContextDecorator, suppress
 from collections import defaultdict, namedtuple
@@ -313,4 +315,32 @@ def strip_key_prefix(prefix, value):
 		else:
 			d[k] = v
 	return d
+
+def timespan_from_outer_bounds(begin=None, end=None):
+	'''
+	Return a `TimeSpan` based on the (optional) `begin` and `end` date strings.
+
+	If both `begin` and `end` are `None`, returns `None`.
+	'''
+	if begin or end:
+		ts = model.TimeSpan(ident='')
+		if begin is not None:
+			try:
+				if not isinstance(begin, datetime.datetime):
+					begin = dateutil.parser.parse(begin)
+				begin = begin.strftime("%Y-%m-%dT%H:%M:%SZ")
+				ts.begin_of_the_begin = begin
+			except ValueError:
+				print(f'*** failed to parse begin date: {begin}')
+				raise
+		if end is not None:
+			try:
+				if not isinstance(end, datetime.datetime):
+					end = dateutil.parser.parse(end)
+				end = end.strftime("%Y-%m-%dT%H:%M:%SZ")
+				ts.end_of_the_end = end
+			except ValueError:
+				print(f'*** failed to parse end date: {end}')
+		return ts
+	return None
